added efs cache construct and cleaned up the main stack
